requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.21.6
Pillow==9.1.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
torch==1.11.0
transliterate==1.10.2
typing_extensions==4.2.0
urllib3==1.26.9
